remove sbom.writer bytes call and consolidate helpers to options pkg
